refactor: merge property-path/transitive-path COUNT ops; cover overlay
Combine fast_property_path_plus_count_all and fast_transitive_path_plus_count_all
into fast_path_plus_count_all — both build a PSOT(p) adjacency map and count
reachable endpoints, so they share a module and the cursor-flavored fast-path
gate.

Hoist the strategy-(b) gate to fast_path_common::allow_cursor_fast_path (one
copy, also reused by fast_group_count_firsts) and switch property-path+ COUNT
off fast_path_store onto it. fast_path_store bails the instant an uncommitted
overlay is present, which left the property-path operator's overlay-aware PSOT
cursor unreachable; it now merges novelty like the transitive operator already
did. Behavior change: <S> <p>+ ?o COUNT(*) gains overlay (and to_t<max_t)
coverage. Add indexed_overlay_property_path_plus_count_reflects_overlay covering
both an overlay assertion and retraction.
